BJS2-40482-Запрос на менторство. Убрал дубликаты RejectionDto и UserService; Объединил тесты для UserService; Подправил основную структуру и структуру папок тестов для MentorshipRequest.
