+++ src/main/java/school/faang/user_service/controller/mentorshiprequest/MentorshipRequestController.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.controller;
+package school.faang.user_service.controller.mentorshiprequest;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -9,9 +9,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import school.faang.user_service.dto.mentorshiprequest.MentorshipRequestDto;
-import school.faang.user_service.dto.mentorshiprequest.RejectionDto;
+import school.faang.user_service.dto.RejectionDto;
 import school.faang.user_service.dto.mentorshiprequest.RequestFilterDto;
-import school.faang.user_service.service.abstracts.MentorshipRequestService;
+import school.faang.user_service.service.mentorshiprequest.MentorshipRequestService;
 
 import java.util.List;
