--- src/main/java/school/faang/user_service/dto/mentorshiprequest/RejectionDto.java
@@ -1,16 +1,0 @@
-package school.faang.user_service.dto.mentorshiprequest;
-
-import jakarta.validation.constraints.NotNull;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class RejectionDto {
-    @NotNull
-    private String reason;
-}
