+++ src/main/java/school/faang/user_service/filters/mentorshiprequest/MentorshipRequestDescriptionFilterImpl.java
@@ -1,23 +1,22 @@
-package school.faang.user_service.filters.impl;
+package school.faang.user_service.filters.mentorshiprequest;
 
 import org.springframework.stereotype.Component;
 import school.faang.user_service.dto.mentorshiprequest.RequestFilterDto;
 import school.faang.user_service.entity.MentorshipRequest;
-import school.faang.user_service.filters.abstracts.MentorshipRequestFilter;
 
 import java.util.stream.Stream;
 
 @Component
-public class MentorshipRequestRequesterFilterImpl implements MentorshipRequestFilter {
+public class MentorshipRequestDescriptionFilterImpl implements MentorshipRequestFilter {
     @Override
     public boolean isApplicable(RequestFilterDto requestFilterDto) {
-        return requestFilterDto.getRequesterId() != null;
+        return requestFilterDto.getDescription() != null;
     }
 
     @Override
     public Stream<MentorshipRequest> apply(Stream<MentorshipRequest> mentorshipRequestStream,
                                            RequestFilterDto requestFilterDto) {
         return mentorshipRequestStream.filter(mentorshipRequest -> mentorshipRequest
-                .getRequester().getId().equals(requestFilterDto.getRequesterId()));
+                .getDescription().contains(requestFilterDto.getDescription()));
     }
 }
