--- src/main/java/school/faang/user_service/service/abstracts/UserService.java
@@ -1,5 +1,0 @@
-package school.faang.user_service.service.abstracts;
-
-public interface UserService {
-    boolean existsById(Long userId);
-}
