--- src/main/java/school/faang/user_service/service/impl/UserServiceImpl.java
@@ -1,17 +1,0 @@
-package school.faang.user_service.service.impl;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import school.faang.user_service.repository.UserRepository;
-import school.faang.user_service.service.abstracts.UserService;
-
-@Service
-@RequiredArgsConstructor
-public class UserServiceImpl implements UserService {
-    private final UserRepository userRepository;
-
-    @Override
-    public boolean existsById(Long userId) {
-        return userRepository.existsById(userId);
-    }
-}
