+++ src/main/java/school/faang/user_service/service/mentorshiprequest/MentorshipRequestServiceImpl.java
@@ -1,18 +1,17 @@
-package school.faang.user_service.service.impl;
+package school.faang.user_service.service.mentorshiprequest;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import school.faang.user_service.dto.mentorshiprequest.MentorshipRequestDto;
-import school.faang.user_service.dto.mentorshiprequest.RejectionDto;
+import school.faang.user_service.dto.RejectionDto;
 import school.faang.user_service.dto.mentorshiprequest.RequestFilterDto;
 import school.faang.user_service.entity.MentorshipRequest;
 import school.faang.user_service.entity.RequestStatus;
 import school.faang.user_service.entity.User;
-import school.faang.user_service.filters.abstracts.MentorshipRequestFilter;
+import school.faang.user_service.filters.mentorshiprequest.MentorshipRequestFilter;
 import school.faang.user_service.mapper.MentorshipRequestMapper;
 import school.faang.user_service.repository.mentorship.MentorshipRequestRepository;
-import school.faang.user_service.service.abstracts.MentorshipRequestService;
-import school.faang.user_service.validator.abstracts.MentorshipRequestValidator;
+import school.faang.user_service.validator.mentorshiprequest.MentorshipRequestValidator;
 
 import java.util.List;
 import java.util.stream.Stream;
