+++ src/main/java/school/faang/user_service/service/user/UserService.java
@@ -13,9 +13,12 @@ import school.faang.user_service.dto.UserDto;
 import school.faang.user_service.dto.user.UpdateUsersRankDto;
 import school.faang.user_service.entity.User;
 import school.faang.user_service.entity.UserProfilePic;
+import school.faang.user_service.entity.UserSkillGuarantee;
+import school.faang.user_service.entity.recommendation.SkillOffer;
 import school.faang.user_service.exception.DataValidationException;
 import school.faang.user_service.mapper.UserMapper;
 import school.faang.user_service.repository.UserRepository;
+import school.faang.user_service.repository.UserSkillGuaranteeRepository;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
@@ -36,6 +39,7 @@ public class UserService {
     private final UserContext userContext;
     private final AvatarService avatarService;
     private final UserMapper userMapper;
+    private final UserSkillGuaranteeRepository userSkillGuaranteeRepository;
 
     public Optional<User> findById(long userId) {
         return userRepository.findById(userId);
@@ -110,5 +114,17 @@ public class UserService {
         return users.stream()
                 .map(userMapper::toDto)
                 .toList();
+    }
+
+    public boolean existsById(Long userId) {
+        return userRepository.existsById(userId);
+    }
+
+    public UserSkillGuarantee addGuaranty(long userId, SkillOffer skillOffer) {
+        UserSkillGuarantee guarantee = UserSkillGuarantee.builder().user(
+                        userRepository.findById(userId).get()
+                ).guarantor(skillOffer.getRecommendation().getAuthor())
+                .build();
+        return userSkillGuaranteeRepository.save(guarantee);
     }
 }
