+++ src/test/java/school/faang/user_service/service/user/UserServiceTest.java
@@ -7,15 +7,19 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
+import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
 import school.faang.user_service.config.context.UserContext;
 import school.faang.user_service.dto.user.UpdateUsersRankDto;
 import school.faang.user_service.entity.User;
+import school.faang.user_service.exception.DataValidationException;
+import school.faang.user_service.mapper.UserMapperImpl;
 import school.faang.user_service.repository.UserRepository;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
@@ -33,17 +37,19 @@ import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
 public class UserServiceTest {
+    private UpdateUsersRankDto usersRankDto;
+    private User user;
+    private List<User> users;
+
     @InjectMocks
     private UserService userService;
 
+    @Spy
+    private UserMapperImpl userMapper;
     @Mock
     private UserRepository userRepository;
-
     @Mock
     private EntityManager entityManager;
-
-    private UpdateUsersRankDto usersRankDto;
-
     @Mock
     private UserContext userContext;
     @Mock
@@ -51,6 +57,16 @@ public class UserServiceTest {
 
     @BeforeEach
     void setUp() {
+        var firstUser = User.builder().id(1L).build();
+        var secondUser = User.builder().id(2L).build();
+        var thirdUser = User.builder()
+                .id(3L)
+                .mentors(List.of(secondUser))
+                .mentees(List.of(firstUser))
+                .build();
+        user = thirdUser;
+        users = List.of(thirdUser);
+
         usersRankDto = UpdateUsersRankDto.builder()
                 .halfUserRank(50.0)
                 .maximumUserRating(100.0)
@@ -142,5 +158,46 @@ public class UserServiceTest {
         assertEquals("User not found", exception.getMessage());
         verify(avatarService, never()).generateRandomAvatar(anyString(), anyString());
         verify(userRepository, never()).save(any(User.class));
+    }
+
+    @Test
+    void testToGetUserDtoById_ShouldReturnCorrectDto() {
+        when(userRepository.findById(3L))
+                .thenReturn(Optional.of(user));
+
+        var userDto = userService.getUserDtoById(3L);
+
+        verify(userMapper, times(1)).toDto(user);
+        assertEquals(user.getId(), userDto.getId());
+        assertNotNull(userDto);
+    }
+
+    @Test
+    void testToGetUserDtoById_ShouldThrowException() {
+        when(userRepository.findById(user.getId()))
+                .thenReturn(Optional.empty());
+
+        assertThrows(DataValidationException.class, () -> userService.getUserDtoById(user.getId()));
+    }
+
+    @Test
+    void testToGetUserDtosByIds_ShouldReturnCorrectDtos() {
+        var ids = List.of(3L);
+        when(userRepository.findAllByIds(ids))
+                .thenReturn(Optional.of(users));
+
+        var userDtos = userService.getUserDtosByIds(ids);
+
+        assertNotNull(userDtos);
+        assertEquals(users.size(), userDtos.size());
+    }
+
+    @Test
+    void testToGetUserDtosByIds_ShouldThrowException() {
+        var ids = List.of(1L, 2L, 3L);
+        when(userRepository.findAllByIds(ids))
+                .thenReturn(Optional.empty());
+
+        assertThrows(DataValidationException.class, () -> userService.getUserDtosByIds(ids));
     }
 }
